workspace-taizi/deployments/hunyuan-ocr-docker/requirements.txt
# HunyuanOCR 本地开发依赖
# pip install -r requirements.txt

# OCR
# hunyuan-ocr  # 官方包，待发布后启用

# API Server
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# 模型 & 数据
huggingface_hub==0.23.0
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python-headless>=4.9.0
requests>=2.31.0

# 日志 & 工具
loguru==0.7.2
python-dotenv==1.0.1

# 测试
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0

# Lint (CI用)
flake8==7.0.0
black==24.4.2
isort==5.13.2
mypy==1.10.0
